requirements.txt
# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# PySpark and Lakehouse Frameworks
pyspark==3.5.0
delta-spark==3.0.0
pyiceberg==0.6.1

# Data Processing
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.3

# ============================================================================
# CONFIGURATION & CLI
# ============================================================================

python-dotenv==1.0.1
click==8.1.7
pyyaml==6.0.1

# ============================================================================
# LOGGING & MONITORING
# ============================================================================

structlog==24.1.0
colorama==0.4.6

# ============================================================================
# UTILITIES
# ============================================================================

tqdm==4.66.1
tabulate==0.9.0
jinja2==3.1.3

# ============================================================================
# TESTING (Optional)
# ============================================================================

pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# ============================================================================
# CODE QUALITY (Optional)
# ============================================================================

black==24.1.1
flake8==7.0.0
mypy==1.8.0
